Fix position after paredit-wrap-round
Original paredit leaves cursor just after opening parenthesis, which is
extremely useful, as this is essentially where the cursor was before
pressing the key binding module the opening parenthesis. It is also
exactly the position, where you'd add a function name to call on
whatever was just wrapped in parentheses.